Revises some tgui implimentation
Gets rid of the beaker tracking entirely from Chem master's buffer. Because... it's the buffer. BeakerContents.tx -> ChemicalContents.tsx to reflect that it controls both now.
