+++ tgui/packages/tgui/interfaces/ChemDispenser/ChemDispenserBeaker.tsx
@@ -1,7 +1,7 @@
 import { useBackend } from 'tgui/backend';
 import { Box, Button, Section, Stack } from 'tgui-core/components';
 
-import { BeakerContents } from '../common/BeakerContents';
+import { BeakerContents } from '../common/ChemicalContents';
 import { removeAmounts } from './constants';
 import type { Data } from './types';
