+++ tgui/packages/tgui/interfaces/ChemMaster/index.tsx
@@ -33,13 +33,10 @@ export const ChemMaster = (props) => {
         <ChemMasterBeaker
           beaker={beaker}
           beakerReagents={beaker_reagents}
-          bufferReagents={buffer_reagents}
           bufferNonEmpty={buffer_reagents.length > 0}
         />
         <ChemMasterBuffer
-          beaker={beaker}
           mode={mode}
-          beakerReagents={beaker_reagents}
           bufferReagents={buffer_reagents}
         />
         <ChemMasterProduction
