+++ tgui/packages/tgui/interfaces/common/ChemicalContents.tsx
@@ -27,3 +27,24 @@ export const BeakerContents = (props) => {
     </Box>
   );
 };
+/** Like above, but for buffer contents */
+export const BufferContents = (props) => {
+  const { bufferContents = [], buttons } = props;
+  return (
+    <Box>
+      {(bufferContents.length === 0 && (
+          <Box color="label">The buffer is empty.</Box>
+        ))}
+      {bufferContents.map((chemical, i) => (
+        <Box key={i} width="100%">
+          <Stack align="center" justify="space-between">
+            <Stack.Item color="label">
+              {formatUnits(chemical.volume)} of {chemical.name}
+            </Stack.Item>
+            {!!buttons && <Stack.Item>{buttons(chemical, i)}</Stack.Item>}
+          </Stack>
+        </Box>
+      ))}
+    </Box>
+  );
+};
